Add an AutoEditor that updates according to the class
This hides or shows the appropriate editors based on the declared
classes that the editor can edit.

The implementation is very naieve at the moment, in that the possible
editors are hard coded, and each is instantiated and remains connected
to the EditorController regardless of whether they are able to edit
the object or not.

Improvements should include instantiating editors on an as-needed
basis, and disconnecting them from the EditorController if the object
is no longer of a class which the editor can display/edit.

This also introduces SSL/TLS handling based on SNI (NOT the requested
hostname in the SOCKS request - that remains to be implemented as a
modification to the SniHandler), along with the necessary boiler-
plate to facilitate loading the CA keypair/certificate, and generating
certificates based on the requested name.