Major updates to the Graph, including:
Making the servers start/stop/restart when nodes are created, modified
or deleted. This makes the entire graph that much more dynamic,
avoiding the need to save and reload the graph to make it work.

Also added warning overlays to the graph, by surrounding handlers
with Exception catchers. These annotate the graph to show where the
exceptions were thrown from, hopefully much more useful than just
showing them in the InterceptPanel. (This feature seems to have been
broken in the process, unfortunately!)