Remove the ExceptionCatcher following the TargetSpecific handler
It's probably not critical that we do, and not ideal that 
TargetSpecificChannelHandler is aware of the ExceptionCatcher.

However, if we do not remove it, the ExceptionCatcher added after the
TargetSpecificChannelHandler ends up getting shunted to the end of the 
pipeline, after any handlers added at the option selection stage of this 
handler. It shouldn't ever get reached, but if it does, it will give an
odd result.

Let's avoid that.