Rework ObjectMapper to not reuse Encoders/Decoders
Decoders change internal state, so we cannot count on a pipeline
remaining usable across multiple invocations.

Store a list of classes that can be creatd using .newInstance(), and
construct a new Pipeline for each mapping that needs to be done.

It feels wasteful, but when we can't be sure what context is left behind
from the previous object, there seems to be no alternative.