Introduce a SubChannel, and SubChannelHandler
The purpose of the SubChannel is to provide a "self-contained" Channel
for user handlers to be added to, such that there will always remain an
internal handler at the end of the pipeline.

This is achieved by directing incoming events on the primary channel
into the SubChannel via the SubChannelHandler, through the user
handlers, then potentially back into the primary channel. The main
advantage of this scheme is to ensure that events will be logged even
when the pipeline contains IndeterminateChannelHandlers, where the
InterceptHandler is only added to the pipeline AFTER events like the
ConnectTarget user event have been fired.

In the normal scheme of things, if a RelayHandler is added to the
SubChannel, no events should reach the end of the primary channel,
because RelayHandler should only forward appropriate events into the
other SubChannel, and not pass them further up its own pipeline.