Update SimpleBinaryModificationHandler to make changes to a copy
Changing the original ByteBuf means that any other places that hold
references to it will also reflect the changes. This is generally not
desirable, we typically want to see a before and after.

Now we make a copy only if there are changes to be made, and make the
changes to that copy.