Implement support for Jython ChannelHandlers
The main requirement is to assign the ChannelHandler instance to _,
so that it can be retrieved. Unfortunately, Jython differs from
Groovy in that it does not return the last expression if the script
contains statements as well (as far as I understand it).

Here is an example Jython ChannelHandler implementation:

from io.netty.channel import ChannelDuplexHandler, ChannelHandler
from io.netty.channel import ChannelHandlerContext
from io.netty.channel import ChannelPromise

class JythonHandler(ChannelDuplexHandler, ChannelHandler):
    def __init__(self):
        print "Init"
        pass

    def channelRead(self, ctx, msg):
        ctx.fireUserEventTriggered("Hello from Jython")
        ctx.fireChannelRead(msg)

_=JythonHandler()